requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
boto3>=1.26.0
pyyaml>=6.0.0
botocore>=1.29.0
